+++ src/integrations/next-client.tsx
@@ -1,5 +1,8 @@
 'use client';
 
+export { AutoTranslate } from './react-auto-translate';
+export type { AutoTranslateProps } from './react-auto-translate';
+
 import {
   createContext,
   useContext,